renderer component and :art:
